+++ RELEASES.md
@@ -14,6 +14,10 @@
 
 - fixed a bug related to missing handling for `ActionState::consumed`
 
+### Usability
+
+- exported `ActionState::action_data_mut_or_default()`
+
 ## Version 0.13.0
 
 ### Breaking Changes